requirements.txt
fastapi
uvicorn
httpx
telethon
python-dotenv
cleantext
websockets
dateparser